tests/wpt/tests/css/css-values/media-progress-serialize.tentative.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-values-5/#progress">
<link rel="author" title="sakhapov@chromuim.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../support/serialize-testcommon.js"></script>
<div id=target></div>
<style>
:root {
  font-size: 10px;
}
#target {
  font-size: 10px;
}
</style>
<script>
function test_serialization(t,s,c) {
    test_specified_serialization('opacity', t, s);
    test_specified_serialization('transform', `scale(${t})`, `scale(${s})`);
    test_computed_serialization('opacity', t, c);
    test_computed_serialization('transform', `scale(${t})`, `matrix(${c}, 0, 0, ${c}, 0, 0)`);
}

test_serialization(
    'media-progress(width, 0px, 1000px)',
    'media-progress(width, 0px, 1000px)',
    '0.8',
);
test_serialization(
    'calc(0.1 * media-progress(height, 0px, 10em))',
    'calc(0.1 * media-progress(height, 0px, 10em))',
    '0.6',
);
test_serialization(
    'calc(media-progress(width, 0px, 100px) / 20)',
    'calc(0.05 * media-progress(width, 0px, 100px))',
    '0.4'
);
test_computed_serialization(
    'width',
    'calc(1px * media-progress(height, abs(11em - 10rem), 110px / 10))',
    '590px',
);
</script>
